Fix: Migrate husky hooks to v9 format (#701)
* Add husky prepare script

Runs automatically on npm install

Creates .husky/_ directory (which is gitignored, hence needs this script to auto init on install)

* Move husky pre-commit hook to .husky/

package.json field not compatible with Husky v9 (was a v4 feature)

* Amend lint-staged tasks

Split entries to avoid race conditions
(https://github.com/lint-staged/lint-staged?tab=readme-ov-file#task-concurrency)

* Move test script out of lint-staged to own action

We want to run the whole test script if a file is change, not only on
changed test files. Else we could introduce a failing change but if the
right test file isn't staged, the test won't be run.